Merged PR 36319: Check previous saved type when saving image types config
## What's being changed

Currently In order to identify if image type changed we compare the current value with the "string value" of the old one. We added an extra condition to check if the type value is string or no.

## Why it's being changed

For a few merchants previous logic was erroring because of some reason the type of old value wasn't the expected.

## How to review / test this change

- Go to Stores > Configuration > Dotdigital > Configuration and find the "Image Types" section (there are 4 dropdowns there)
- Save the image type configurations and confirm that you get no errors.
- Confirm that we identify correctly if value changed or not
- Confirm that value is getting saved as expected

Related work items: #170230