Merged PR 36456: Prevent undefined index error in Resubscriber sync
## What's being changed

We're adding some extra checks to protect against mismatches between email addresses stored in Dotdigital and Magento.

## Why it's being changed

Evidently it _is_ possible for email addresses to be stored in DD and not converted into lowercase. We needed to convert to lowercase wherever we handle email addresses here to avoid inconsistencies.

In addition, we have a fail-safe `isset()` check for the index before using it.

## How to review / test this change

- Subscribe chaz@emailsim.io in Magento
- Sync to Dotdigital
- Modify email address in DD to have a capital C
- Unsubscribe in Magento (this can be done by modifying the newsletter_subscriber table - change subscriber_status to 3 and add 5 minutes to change_status_at. Don't re-sync or run the importer at this stage.)
- Run the resubscribe cron. Easiest way to do this is to comment out LL.120-121 in Newsletter\Subscriber.
- Do not see an 'undefined index' error.

Related work items: #170563