Merged PR 36631: Resolve importer error if contact sync cron executes every 60'
## What's being changed

In this PR we return instantly that the cron duration is 60 mins in `getDecodedCronValue()`.

## Why it's being changed

This function is not getting called only on configuration save (where we initialize the cron patterns with the offset) but in importer cron too, to figure out the last contact sync cron time.

https://github.com/dotmailer/dotmailer-magento2-extension/blob/master/Model/Sync/Importer/ImporterReportHandler.php#L289

In case that the cron schedule for contact sync is set to run every 60' (pattern is X * * * * *) the decoder function throws an exception instead of sending the correct value.

## How to review / test this change

_Test report handler_
- set contact cron to run every 60'
- make sure that you have in your email_importer bulk contact records
- Run importer sync twice
- Make sure that in second run you get the correct value [here](https://github.com/dotmailer/dotmailer-magento2-extension/blob/master/Model/Sync/Importer/ImporterReportHandler.php#L289)

_Test config changes_
- save different values for cron timings
- test saving both 15/30 minutes AND 60 minutes

Related work items: #171487