Merged PR 36880: Automate the integration setup process
## What's being changed

We've introduced a 'Setup Integration' button in Dotdigital > Accounts. When clicked, this:
- enables select syncs
- maps address books
- maps data fields
- enables easy email capture for newsletter
- sends 10 orders (and their contacts) to Dotdigital
- sends 10 products to Dotdigital
- checks ddg_automation cron jobs are scheduled in the next hour

In addition this PR adds an account status bar to indicate the Dotdigital account connected in the current scope.

## Why it's being changed

This routine will speed up the process of connecting with Dotdigital. It automates some admin steps commonly taken by our support teams when onboarding merchants, and by pre-loading (i.e. instantly, without waiting for cron jobs) real order, customer and product data in Dotdigital, we enable access to some key areas of functionality in the platform.

## How to review / test this change

- Go to **Dotdigital > Accounts**
- Review the new account status bar
- Click Set Up Integration
- Review the steps that are revealed in the UI
- More test steps outlined on component PRs

Related work items: #170832, #171297