Merged PR 37424: Optimize account checks in integration setup
## What's being changed

We have added local properties to store connection state and account information to the `AccountDetails` class to avoid quiring dotdigital more than required., we have optimized a few methods in the `SetupIntegration` class  to insure the reported data is accurate and set correctly.

## Why it's being changed

We are calling the api several time during  a single setup integration run and not returning a false if the integration step failed several steps.

## How to review / test this change

- Create a new trial account
- Run setup integration
- Confirm the setup integration functions as expected.

## Notes

Follow the go bag go test steps to be 100% sure.

Related work items: #175030, #175031