Merged PR 37442: Fix dummy records on website scope
## What's being changed

We added an extra check when saving API configurations in website level to check if email is declared due to dummy cart insight data sending.

## Why it's being changed

Currently as it is email is not specified in website scope when attempting to send cartInsight dummy data on account saving configurations.

## How to review / test this change

- Go to Stores > Configuration > Dotdigital > Accounts
- Switch to website level
- Add a valid EC account credentials
- Hit Save
- Confirm that in EC a new contact created (with the account owner's email) and dummy cartInsight data collection assigned in to it

Related work items: #175347