Merged PR 38242: Add parent name to fallback item schema
## What's being changed

This PR adds a `parent_name` to the fallback item schema.

## Why it's being changed

We never used to include this, but there's no reason not to since we have the data in the `sales_order_item` table. For an order with these line items:
```
configurable (can't find product)
simple (can't find product)
```
the fallback data (for the simple product - nothing is output for the configurable parent) should indicate the line item parent name.

## How to review / test this change

- Place an order for a configurable product
- In `sales_order_item`, change both line items to have an invalid product_id
- Run order sync
- Note the parent_name in the item schema
- Reset your order data and re-sync everything, checking the logs for any 'undefined' errors or similar.

Related work items: #178941