Merged PR 38275: Exclude Create In customer attribute from mapping
## What's being changed

In this PR we added "Create In" (store_id code) customer attribute in our exclusion list.

## Why it's being changed

As it stands, if this attribute is selected, then no value is synced in to EC (our code doesn't handle the nested array value structure). However, because this value maps the store view name, and we already provide data field options for this attribute, it's better to exclude this attribute (in the same way as we are excluding `website_id`).

## How to review / test this change

- Before checkout to this branch go to Stores > Configuration > Dotdigital > Data mapping
- In Custom Attributes section confirm that you see "Create In" option in the Attribute dropdown
- Checkout to the branch and confirm that is now disappeared.

## Notes
- We spotted this because of a support ticket (see related work items).
- "Create In" looks like a default customer attribute form magento although is not defined here https://docs.magento.com/user-guide/stores/attributes-customer.html#default-customer-attributes
- Inspect "eav_attribute" table and check if you get this value by default. (Note we offer only attributes where entity_type_id = 1).
- In my case in Stores > Attributes > Customers i don't see this attribute

Related work items: #178273, #178715