Merged PR 38415: Allow multiple contains/not contains conditions in ANY exclusion rules
## What's being changed

Exclusion rules logic.

## Why it's being changed

Multiple conditions on the same column are only allowed in ANY mode, and in this case, conditions are passed to `addFieldToFilter` as an array, resulting in `OR` in MySQL. However, if the conditions have 'contains', or 'not contains', `OR` doesn't actually work.

![Screenshot 2022-07-08 at 11.47.57.png](https://dev.azure.com/dotdigital/b2ba4bc1-4a9a-48f7-a69a-5cf84134b7c4/_apis/git/repositories/cc22fc8a-89ce-4edf-ba9c-81a01b55e11a/pullRequests/38415/attachments/Screenshot%202022-07-08%20at%2011.47.57.png)

This currently translates into this SQL:
```
... AND ((`main_table`.`customer_email` NOT LIKE '%emailsim.io%') OR (`main_table`.`customer_email` NOT LIKE '%dotdigital.com%') OR (`main_table`.`customer_email` NOT LIKE '%chaz.com%'))
```
where, even if the quote email address is chaz@emailsim.io, it won't be excluded because the other OR conditions remain true.

This PR updates the logic to add separate filters for each contains or not contains (i.e. 'like' or 'nlike') condition in an ANY scenario. Now the logic will be:
```
... AND (`main_table`.`customer_email` NOT LIKE '%emailsim.io%') AND (`main_table`.`customer_email` NOT LIKE '%dotdigital.com%') AND (`main_table`.`customer_email` NOT LIKE '%chaz.com%')
```

Note that for 'is', 'is not' or 'is empty' conditions, we'll still pass all of those as an array so that logic is not changed.

## How to review / test this change

- Construct an ANY type exclusion rule for either Abandoned Cart or Reviews
- Add multiple 'contains' conditions for Email (or any other column that allows this type of condition)
- Prepare a quote or an order that you expect to be excluded according to the conditions
- Run the AbandonedCart or Review sync CLI
- Confirm your quote or order is excluded
- Repeat with the opposite logic (i.e. not contains) and see your quote / order processed as normal
- Add additional conditions for other columns e.g. product attributes, then re-test
- Switch to an ALL combination rule, re-test (note that in an ALL, any extra conditions for the same column will be ignored)

## Notes

_Testing with reviews_
There is a step in the review sync that sends out review reminder emails to customers who have placed orders.

- Go to Dotdigital > Automation > Review settings and set up review reminder emails for Pending orders
- Place an order for a product
- Now run review sync. You might need to manually override the from and to values in Model\Sales\Order::searchOrdersForReview().
- Exclusion rules in this scenario will prevent review reminders being sent where the order's properties match the conditions in the rule.

Related work items: #176532