Merged PR 39012: Fix integration tests failure
**What Changed**
We are now using the `getLastItem()` instead of `getFirstItem()` in order to assert collection items in this specific integration test.

**Why Changed**
Because when we running the whole suite there might be still already records exist in our tables from previous tests. As a result we pick the wrong item to assert.

**How to review it**
Make sure that integration tests pass

Related work items: #181973