Merged PR 38999: Check for status property in ImporterProgressHandler
## What's being changed

We've added an extra check for the response `status` property when looking up in-progress imports.

## Why it's being changed

In rare cases it is possible that we make calls to either of these endpoints:
- https://developer.dotdigital.com/reference/get-transactional-data-import-status
- https://developer.dotdigital.com/reference/get-contact-import-status

and what we get back has neither a `message` (i.e. a failure) or a `status` key. This could happen in the case of a local network issue for example. So we can never 100% guarantee that such properties will be present, and therefore should always check with `isset()`.

## How to review / test this change

- Queue a two-stage import by running e.g. customer sync, catalog sync
- Run the importer
- Check the queued import now has status 1
- Re-run the importer
- Confirm that the import switches to status 2
- Try to replicate the network issue by e.g. going offline and re-running the test

Related work items: #182139, #182266