Merged PR 39025: Fix cart insight image for configurable variant if parent has customizable options
## What's being changed

ImageFinder - specifically the bit of logic that finds the variant for a configurable quote item, if "Configurable Product Option" is set to "Itself".

## Why it's being changed

We were using the SKU to get the child product id, but this breaks where customizable options are set on the parent product (resulting in a dynamic SKU).

Ordinarily, quote items for configurable products have the same SKU e.g.
| Quote item | |
| -- | -- |
| Parent | MH02-XS-Black |
| Child | MH02-XS-Black |

However, once customizable options are configured, we have e.g.
| Quote item | |
| -- | -- |
| Parent | MH02-XS-Black-custom-sku-x |
| Child | MH02-XS-Black |

This was causing problems in cart insight data, because we were trying to find a product for the dynamic SKU, not finding one, and falling back to the parent product (which might have a different thumbnail).

## How to review / test this change

- On a configurable product, set up a Customizable Option in the section on the product admin page
- Verify the option is displayed on the storefront
- Add a variant of this configurable product to your cart (the image of the simple product should be different from the image of the parent)
- In Stores > Configuration > Sales > Checkout > Shopping Cart set 'Configurable Product Image' to 'Product Image Itself'
- In the storefront shopping cart, the image you see should be the child's image
- Run AbandonedCart sync
- Confirm that the line item thumbnail is the correct one belonging to the child
- Set 'Configurable Product Image' to 'Parent Product Thumbnail' and retest
- Confirm that the line item thumbnail is the correct one belonging to the parent

## Notes

We _could_ amend `CartInsight\Data::loadProduct()` to use my code i.e.
```
$variants = $item->getChildren();
$this->productRepository->getById($variants[0]->getProductId());
```
which would work, and produces an identical result. And it feels like it makes more sense as well. Instead of trying to get the child by SKU, failing, and falling back to the parent product, we would go straight to the child. However, since it already works (thanks to `Magento\ConfigurableProduct\Model\Product\Type\Configurable::getPrice()` being smart enough to get the child’s price), I’m inclined to not touch it.

Related work items: #182094