Merged PR 39041: If change_status_at is empty, always resubscribe
## What's being changed

Resubscriber sync. We're preventing any errors that occur if we don't find a `change_status_at` value for the matching Magento subscriber, and we're resubscribing as expected.

## Why it's being changed

It emerged from this Github issue: https://github.com/dotmailer/dotmailer-magento2-extension/issues/588.

The Resubscriber class:
- goes to Dotdigital to find recently modified contacts
- filters those for contacts subscribed in the time window (i.e. since the last cron run)
- finds matching subscribers in newsletter_subscriber who did not change status more recently
- performs resubscribes for these contacts in Magento

If for some reason the matching subscriber had an empty `change_status_at`, we would not resubscribe - but now we do. We treat such subscribers in the same way as if they had a `change_status_at` value and it was older than the Dotdigital updated at.

## How to review / test this change

- Subscribe an email address and sync them into Dotdigital
- Now, unsubscribe in Magento
- Manually clear the change_status_at value in newsletter_subscriber
- Run trigger the Resubscribe sync (instructions here in the notes https://github.com/dotmailer/dotmailer-magento2-extension/commit/c691960634ad97bdfefb1cb34c6dd5534a1820e1, or do it from the admin)
- Ensure there is no datetime-related error
- Ensure the resubscribe happens as intended

Related work items: #182468