Merged PR 39201: Fix unit test flagged as risky
## What's being changed

The ServerSentEventTest unit test.

## Why it's being changed

The amended test opens a stream and checks various properties of the stream (it's actually an integration test since it doesn't use mocks). It generated a warning:
```
1) Dotdigitalgroup\Email\Test\Unit\ServerSentEvent\ServerSentEventTest::testCreateResponse
Test code or tested code did not (only) close its own output buffers

OK, but incomplete, skipped, or risky tests!
```

PHPUnit throws this error if the output buffer level at the end of the test does not match the level at the start. This PR removes the warning by restoring the output buffer level to 1.

## How to review / test this change

- See green (as in really green) build.

Related work items: #183101