Merged PR 39633: Fix UpdateApiEndpointPath data patch
## What's being changed

The UpdateApiEndpointPath data patch.

## Why it's being changed

It is possible for merchants to have both _old_ (connector/api/endpoint) and _new_ (connector_api_credentials/api/endpoint) API config paths at the same time. e.g.

| config_id | scope | scope_id | path | value |
| --- | --- | --- | --- | --- |
| 1 | default | 0 | connector/api/endpoint | https://r1-api.dotmailer.com |
| 2 | default | 0 | connector_api_credentials/api/endpoint | https://r1-api.dotmailer.com |

This can happen after the upgrade to 4.18.0, if a merchant were to re-save account credentials in any given scope. The problem occurs at upgrade to 4.18.2 (when this patch was introduced).

With the data set as above, the patch fails and breaks `setup:upgrade`. This PR fixes the patch to either:
- delete any rows where an updated path already exists for that scope_id
OR
- update the path (as previously)

## How to review / test this change

- Set up a multi-website instance
- Get your endpoint rows into a similar state per the above example i.e. some new, some old, across different scopes
- Delete the UpdateApiEndpointPath row from the `patch_list` table.
- Run `setup:upgrade`
- Ensure there is no error thrown
- Confirm that old paths are either removed or updated

Related work items: #184909