Merged PR 39861: Cast explode param to string in getStreet
## What's being changed

The class that collects data for Order export.

## Why it's being changed

Street is actually a required field in Magento front end, but I guess not all front ends would be consistent, and orders can be imported from other systems - so it is possible for the 'street' address key to be null. This fixes the PHP8.1 exception that is thrown if that is the case.

## How to review / test this change

- Hack your order billing address or shipping address to remove the street
- Clear cache
- Run order sync
- See the process complete without error, including your order

Related work items: #185947