Merged PR 40415: Use key storage on quoteItems property
## What's being changed

Abandoned cart automation processing. We had some code to store a quote's items in memory, and we're now handling this as an array to store all quote items across a larger set of quotes.

## Why it's being changed

We store quote items in memory here in order to only load each quote's items once. We first check if the quote has any items (and if not we exit the loop for this quote), then we process the items. However in a set of many quotes, the quote items property would be set once, then never reset, hence all related contacts would have the same data fields pushed. This PR fixes the problem.

## How to review / test this change

- Enable only AC automation program enrolment and not abandoned cart campaign series
- For two different contacts drop two different carts (make sure you have different product in each cart)
- Run AC sync to create 2 new AC automation enrolment rows in email_automation table
- Run automation sync
- Inspect the contact records in Dotdigital platform
- Confirm that contact 1 ABANDONED_PROD_NAME matches their quote
- Confirm that contact 2 ABANDONED_PROD_NAME matches their quote

Related work items: #188328