Merged PR 41263: Update does not belong to AbstractDb resource class
## What's being changed

In pendingContactUpdater we are using `$resourceModel->update()` to update function rows and we mark this resource model as an AbstractDb one. But `update()` function is implemented only in Abandoned and Automation resource models and it's not a core function as a result upgrade compatibility tool complains. So in this PR we specify the type of resourceModel as Automation|Abandoned.

## Why it's being changed

To fix the error.

## How to review / test this change

There is nothing really we can test as only docblock updated.

Related work items: #190604