Merged PR 41341: Fix loadAttributeOptions error in upgrade compatibility tool
## What's being changed

We used to instantiate the private property factories for the Type class in the constructor. In the Magento 2 upgrade compatibility tool this was returned as an error because the loadAttributeOption() function doesn't exist in the factory object. Now, we only instantiate the factory properties when we actually need them.

## Why it's being changed

To resolve the error.

## How to review / test this change

- Navigate to Marketing > Exclusion rules click Add new rule go to "Conditions" tab and make sure that conditions loaded as expected.
- Make sure that upgrade compatibility tool returns no error.

Related work items: #190601