Merged PR 41470: Fix type issues in the Consent model
## What's being changed

Some methods in the Consent model.

## Why it's being changed

The main problem was a bug where, if a merchant had not set consent text in Dotdigital > Configuration > Consent, subscriber sync would fail owing to a PHP error on `strlen()`.

In addition to this, we've tightened up return types from a number of methods, and moved a few methods back from the Config Helper into the model.

## How to review / test this change

- Enable consent in Dotdigital > Configuration > Consent, but leave both text boxes empty
- Subscribe an email address on the store front
- Sync subscribers, checking no errors are thrown
- Create a new customer account, sign in, and in the account page check the box to sign up to newsletters
- Confirm that a payload was sent to DD with consent and preference data
- Go back to the Consent options and enter text in both boxes, and repeat your tests, confirming that the right text was sent to DD

Related work items: #188342