Merged PR 41264: Fix loadByCustomerId and loadByCustomerEmail error in upgrade compatibility tool
## What's being changed

The Collection class for dotdigital contacts contains the methods `loadByCustomerId` and `loadByCustomerEmail`. Currently in the Contact class we're using the core (and deprecated) method `getCollection` in order to fetch the collection and to use the methods described. This method by default returns an instance of AbstractDbCollection which our Contact Collection class inherits. However loadByCustomerId and loadByCustomerEmail don't exist currently in AbstractDbCollection and as a result we're getting upgrade compatibility errors.

In order to skip this error we no longer use `getCollection` method and we are calling directly our collection factory which contains directly the implementation of these functions.

We've also removed some unused methods, and updated the name of the ResetContactImport observer to ResetCustomerImport, since it only concerns the resetting of synced customers (not any other type of contact).

## Why it's changed

We made this change in order to resolve the errors we get from magento 2 upgrade compatibility tool.

## How to review / test this change

- Run the [magento2 upgrade compatibility](https://experienceleague.adobe.com/docs/commerce-operations/upgrade-guide/upgrade-compatibility-tool/use-upgrade-compatibility-tool/run.html) tool and make sure that the error has gone.
- Sync customers
- Update any value in the Data Mapping section of the config
- Ensure that all customers are reset

Related work items: #190602, #191917