Merged PR 41266: Fix loadByOrderIdandQuoteId error in upgrade compatibility tool
## What's being changed

The Collection class for dotdigital orders contains the method `loadByOrderIdandCustomerId`. Currently in the Order class we're using the core (and deprecated) method `getCollection` in order to fetch the collection and to use the method described. This method by default returns an instance of AbstractDbCollection which our Order Collection class inherits. However `loadByOrderIdandCustomerId` doesn't exist currently in AbstractDbCollection and as a result we are getting upgrade compatibility errors.

In order to skip this error we no longer use the `getCollection` method and we are calling directly our collection factory which contains directly the implementation of this function.

## Why it's changed

We made this change in order to resolve the errors we get from magento 2 upgrade compatibility tool.

## How to review / test this change

- Run the [magento2 upgrade compatibility](https://experienceleague.adobe.com/docs/commerce-operations/upgrade-guide/upgrade-compatibility-tool/use-upgrade-compatibility-tool/run.html) tool and make sure that the error has gone.
- Place an order on the storefront
- Send an invoice for the order; confirm that the matching row is set to 'processing' in the `email_order` table
- Issue a credit memo for the order; confirm that the matching row is set to 'closed'

Related work items: #190603