Merged PR 41510: Align order items and order process item refactor
## What's being changed

In this PR we added `product_id`,`parent_id`, `price_inc_tax` fields in product details for orderInsight data. Also we've refactored the process of order items  and all this logic moved in to a new class.

## Why it's being changed

- We've added the extra fields in order to align magento order schema with the rest of the dotdigital connectors.
- We've refactored the `processOrderItems()` because previously it was a 150 lines "hard to read" function. Also we've moved in the new class all the private methods that the old `processOrderItems()` used to call. That means Order.php class is half size now.

## How to review / test this change

I will provide a few test scenarios that need to be tested.
Ideally all these scenarios should be tested in develop and in the current branch in order to compare the order schema.

**1st scenario**
- Complete an order including Configurable, Bundle, and Simple product
- Compare schema for both develop and current branch

**2nd scenario**
- For the order in first scenario log in to backend and delete the configurable parent
- Run upgrade and indexer
- Compare the schema in both cases
(why to do this? Checked the removed piece of code in Line 452 in Order.php)

**3rd scenario**
- Go to Stores > Configuration > Dotdigital > Configuration > Transactional Data and enable `Include Product Custom Options In Order Sync`
- Select a few options
- Repeat steps scenario 1

Finally make sure that newly added fields (see description) are populated correctly.

Related work items: #192044