Merged PR 41730: Use POST instead of PUT to add data fields to a contact
## What's being changed

One of our API client methods: `updateContactDatafieldsByEmail`. Historically (since Magento 1) we created a contact first before doing a `PUT` request against a contact id. This PR rolls everything into a single `POST` request.

## Why it's being changed

Our API supports POSTing contacts _with_ data fields, and there's no apparent advantage of checking first before adding the data fields - since the response will be the same in the event that we've not been able to create a contact. The upside with this change is that we omit the extra API request.

## How to review / test this change

- Run any flow that involves a data field update (e.g. automation sync, campaign sync, or unsubscribes as processed by the importer sync)
- Confirm that data fields are still set as expected in DD
- Confirm that an error response or error message is still handled as expected (fx by suppressing the contact DD-side and re-running the sync)

Related work items: #194933