Merged PR 42034: Send empty response if no Feefo reviews are available
## What's being changed

In this PR we modified the Feefo template and we added an initial if statement to check if any reviews are available **before** we print out any HTML content.

## Why it's being changed

Currently our EDCs if their total HTML content chars are < 10 return 204 response. For Feefo EDC the template file was sending back an empty html table with total html content chars > 10. As a result instead of getting 204 response we get a useless empty html table. With this change we prevent that and we get the expected behaviour.

## How to review / test this change

- Go to stores > configuration > dotdigital > automation
- Enable feefo logon by setting the "Feefo Logon" field.
- Run the product reviews edc in the browser and select a quote_id that has no reviews.
- Confirm that you get 204 response instead of an empty html table

Related work items: #193523