Merged PR 42488: Exclude orders with empty customer_email in migration
## What's being changed

In this PR we exclude orders where customer_email is either NULL or an empty string when populating email_order table, and when we inserting guests from orders in email_contact table.

## Why it's being changed
We spotted a few cases that people have orders in sales_order table with null customer_email field. Because of this we can end up with:
- Php error because in case we attempt to pass null arguments in `strcasecmp` during migration scripts run.
- We might attempt to send orderInsight data to nowhere during order sync.

## How to review / test this change

- Update a few orders of your sales_order table with null or empty string customer_email field.
- Ensure that no record created for this specific order in email_order table.
- Ensure that you get no error when running bin/magento dotdigital:migrate

Related work items: #197565