Merged PR 42893: Run Integration Insights sync at account setup using message queues
## What's being changed

Integration insights sync is now queued whenever account credentials are updated (including when a trial account is created).

## Why it's being changed

Although we used to send integration data to Dotdigital at account connection, we moved this to a weekly cron once we needed to send configuration data as well - on the basis that this was too big an operation to do synchronously. Using message queues we can sync the information as close to real time as the technology allows. [Note that the weekly cron remains in place].

Another driver for this change was the need to have integration information in place before a merchant can start using our (forthcoming) product notifications feature.

## How to review / test this change

1. Ensure you run  `bin/magento setup:upgrade` and that your crons are setup
2. Add a breakpoint to `Dotdigitalgroup\Email\Model\Sync\Integration\IntegrationInsights@L39`
3. Navigate to Configuration > Dotdigital > Account
4. Remove your account and save.
5. Add your account again and save. **Ensure the account is valid
6. Run `bin/magento cron:run`
7. PHPstorm should break in the `IntegrationInsights::class` observe the behaviour.
8. Check the connector.log contains the following entry:
8.1 `dotdigital.INFO: Integration insight data sent {"integration":"php8.2.magento-enterprise-edition-2.4.6-beta2.localhost","success":true} []`
9. Check that integration insights within DD has been updated / created
10. Test the integration insights using a new trial account via the Laravel microsite.

Related work items: #193694