Merged PR 43107: Protect against null values being passed to round() in cart insight
## What's being changed

Our Cart Insight data payload.

## Why it's being changed

We're checking that the values we pass to `round()` default to 0 (i.e. are not `null` or otherwise false-y). They shouldn't be null of course, the methods we're using should ALL return a float - but somehow we are seeing this error in merchant logs.

## How to review / test this change

- Drop a cart
- Configure abandoned cart program enrolment
- Edit your program enrolment interval
- Run abandoned cart sync
- Confirm the payload data is as expected
- Test a `null` value by changing e.g. L192 to:
```
'salePrice_incl_tax' => round(null ?: 0, 2),
```

Related work items: #200422