Merged PR 43301: PHP 7 support fixes
- Fixed Exception with no variable declaration (which is PHP 8+)
- Removed spread operator ... in favour of array_merge for 7.2 and 7.3 support
- removed trailing commas on class constructor
- removed property type association for 7.2 and 7.3 support

We have found a few files within our email module do not support php7.2, we have had to update / downgrade the functionality in favour of php7.2 support.

1. setup or use Magento Enterprise or Commerce Magento 2.3.1 - 2.3.5 ( or any magento that support php7.2 )
2. run` php7.2 bin/magento setup:upgrade`
3. run` php7.2 bin/magento setup:di:compile`
4. observe a successful upgrade and compile

The update to DataFieldCollector goes a little further than simply replacing the spread operator with `array_merge` - it moves the merging outside of the foreach loop, since this doesn't need to be run for each new data field.

Related work items: #202195