Merged PR 43257: Resolve or work around dynamic property assignment
## What's being changed

Dynamic property assignment in catalog sync and review sync.

I initially went with a magic `__set()` method in both cases, but abandoned this and simply added elements to a private `$properties` array in each model (Attribute and Review).

## Why it's being changed

In PHP 8.2, dynamic property assignment (`$this->$var` for example) is deprecated and causes script errors.

## How to review / test this change

First, run catalog, review and order syncs in 2.4.6 on the develop branch using PHP 8.1.x. That way you'll have a record of the JSON being output.

_Review sync_
- Place an order
- Place a review of one of the products in your order
- Approve the review
- Run the review sync
- Compare the schema with the original

_Catalog sync_
- Run catalog sync
- Observe that 'attributes' is now output in all cases
- Assign product attributes for sync
- Ensure 'attributes' is populated
- Compare the schema with the original

_Order sync_
- Assign product attributes for sync
- Place an order for a configurable product
- Run the order sync
- Ensure 'product_attributes' and 'child_product_attributes' are populated
- Compare the schema with the original

## Notes

The `Rating` model has been removed. This model required dynamic property assignment in order to have a flexible key for the 'rating code' (which can be 'Rating', but also be 'Quality', 'Value', 'Price' etc.), but the export actually just required an array, and there was no benefit to trying to use an object of a particular type.

Related work items: #200326, #200385