Merged PR 43503: Move consent configuration to own section
## What's being changed

In this PR we changed the consent text fields section position and for first time we allow merchants to fill consent text fields.

## Why it's being changed

As we introduce support for unified contacts in Dotdigital, we have SMS consent as well as Email consent, and it made sense to group these configs under a single section. This accompanies updates to our SMS module.

## How to review / test this change

- Before you checkout make sure you filled a few consent fields in the old consent section in dotdigital > configuration.
- Go to dotdigital > configuration and confirm that a new section called consent created.
- Confirm that the fields are what described in story description.
- Inspect core_config_data and make sure that paths of consent fields got updated as per new section position.
- Make sure that if old consent were stored in website scope we still respect the scope.

Related work items: #202620