Merged PR 43528: Check if customer email exists before proceed with getBasket link
## What's being changed

In this PR we check if `customer_email` exists in the quote before we redirect user to basket after they click the "Take me to Basket" link.

## Why it's being changed

In case that we didn't capture the email address for a guest quote our abandoned cart EDC still can redirect users to a cart that doesn't belong to anyone.

## How to review / test this change

- As a guest drop an item in to cart.
- run the abandoned cart EDC and click "go to Basket" link.
- confirm that you getting redirect to home page.

Related work items: #203841